Update refresh token method

+++ extra/script/entrypoint.sh
@@ -1,14 +1,17 @@
 #!/bin/sh
 
-# Change to backup directory
-cd ./backup
+# Check if the backup directory exists
+if [ -d "./backup" ]; then
+    # Change to backup directory
+    cd ./backup
 
-# Start migrations
-npm run typeorm migration:run -- -d ./src/helper/datasource.ts
+    # Start migrations
+    npm run typeorm migration:run -- -d ./src/helper/datasource.ts
 
-# Remove unused files
-cd ../
-rm -rf ./backup
+    # Remove unused files
+    cd ../
+    rm -rf ./backup
+fi
 
 # Start project
 npm run start:prod
